requirements.txt
pandas
scikit-learn
flask
flask-cors
pyserial
requests
joblib